refactor: create ecobalyse_data module
